Addressed PR comments - moved gpu to virtualization & removed non-error logs
