+++ pkg/virtualization/core/gpu/gpu.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-package host
+package gpu
 
 import (
 	"log"
@@ -10,11 +10,10 @@ import (
 	"github.com/microsoft/wmi/pkg/base/instance"
 	"github.com/microsoft/wmi/pkg/base/query"
 	"github.com/microsoft/wmi/pkg/constant"
-	"github.com/microsoft/wmi/pkg/virtualization/core/gpu"
 )
 
 // GetPartitionableGpuCollection gets all host partitionable GPUs
-func GetPartitionableGpuCollection(whost *host.WmiHost) (partitionablegpucollection gpu.PartitionableGpuCollection, err error) {
+func GetPartitionableGpuCollection(whost *host.WmiHost) (partitionablegpucollection PartitionableGpuCollection, err error) {
 	query := query.NewWmiQuery("Msvm_PartitionableGpu")
 	rasdcollection, err := instance.GetWmiInstancesFromHost(whost, string(constant.Virtualization), query)
 	if err != nil {
@@ -23,12 +22,11 @@ func GetPartitionableGpuCollection(whost *host.WmiHost) (partitionablegpucollect
 	}
 	defer rasdcollection.Close()
 
-	partitionablegpucollection, err = gpu.NewPartitionableGpuCollection(rasdcollection)
+	partitionablegpucollection, err = NewPartitionableGpuCollection(rasdcollection)
 	if err != nil {
 		log.Printf("[WMI] Error getting new PartitionableGpuCollection for rasdcollection [%s] - Error details [%+v]\n", rasdcollection, err)
 		return
 	}
 
-	log.Printf("[WMI] Got partitionable GPU collection [%s]", partitionablegpucollection)
 	return
 }
